refactor: collapse tui direct runtime lifecycle
Covers F0.1/F0.2, F2.1/F2.2/F2.3, F3.1, F4.1, and F6.1 slice 1A.

TUI runtime startup now requires shared runtime client, session activity stream, and prompt activity stream. Direct raw runtime event bridge, ask bridge, raw-engine runtime client helpers, and direct process manager fallback were removed from production CLI code; tests use loopback service adapters.
