+++ cli/app/ui_runtime_adapter.go
@@ -5,7 +5,6 @@ import (
 	"strings"
 
 	"builder/cli/tui"
-	"builder/server/runtime"
 	"builder/shared/clientui"
 
 	tea "github.com/charmbracelet/bubbletea"
@@ -195,14 +194,6 @@ func (a uiRuntimeAdapter) syncConversationFromEngine() tea.Cmd {
 		return nil
 	}
 	return m.requestRuntimeCommittedConversationSync()
-}
-
-func (a uiRuntimeAdapter) handleRuntimeEvent(evt runtime.Event) tea.Cmd {
-	return a.handleProjectedRuntimeEvent(projectRuntimeEvent(evt))
-}
-
-func (a uiRuntimeAdapter) applyChatSnapshot(snapshot runtime.ChatSnapshot) tea.Cmd {
-	return a.applyProjectedChatSnapshot(projectChatSnapshot(snapshot))
 }
 
 func waitAskEvent(ch <-chan askEvent) tea.Cmd {
